JAMES-2521 Move mailbox probes definitions to API
This makes store dependency no more required by CLI
