+++ mailbox/api/src/main/java/org/apache/james/mailbox/probe/MailboxProbe.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.store.probe;
+package org.apache.james.mailbox.probe;
 
 import java.io.InputStream;
 import java.util.Collection;
