+++ server/container/cli/src/test/java/org/apache/james/cli/ServerCmdTest.java
@@ -39,8 +39,8 @@ import org.apache.james.core.quota.QuotaSize;
 import org.apache.james.mailbox.model.MailboxId;
 import org.apache.james.mailbox.model.SerializableQuota;
 import org.apache.james.mailbox.model.SerializableQuotaValue;
-import org.apache.james.mailbox.store.probe.MailboxProbe;
-import org.apache.james.mailbox.store.probe.QuotaProbe;
+import org.apache.james.mailbox.probe.MailboxProbe;
+import org.apache.james.mailbox.probe.QuotaProbe;
 import org.apache.james.probe.DataProbe;
 import org.apache.james.probe.SieveProbe;
 import org.apache.james.rrt.lib.MappingsImpl;
